requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
Werkzeug==3.0.3
reportlab==4.2.2
gunicorn==22.0.0
python-dotenv==1.0.1
mysqlclient==2.2.4
